requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sendgrid
pandas
gspread
google-auth
temporalio==1.*
